Refactor `EditorPlugins`.
